backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
alembic>=1.12.0
psycopg2-binary>=2.9.7
celery>=5.3.4
redis>=5.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6
pdfplumber>=0.10.2
pytesseract>=0.3.10
pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.2
sse-starlette>=1.6.5
flower>=2.0.0
